# Conflicts:
#	SCOOP_Project/target/classes/kr/or/scoop/mapper/ProjectDao.xml

+++ SCOOP_Project/src/main/java/kr/or/scoop/controller/TeamController.java
@@ -24,6 +24,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.LocaleResolver;
 
 import kr.or.scoop.dao.MemberDao;
 import kr.or.scoop.dao.MyIssueDao;
@@ -56,8 +57,12 @@ public class TeamController {
 	
 	@Autowired
 	private TeamService teamservice;
+	
 	@Autowired
 	private PrivateService privateservice;
+	
+	@Autowired
+	private LocaleResolver localeResolver;
 	
 	@RequestMapping(value = "team.do" , method= {RequestMethod.POST,RequestMethod.GET})
 	public String CreateProject(TeamPjt team) {
@@ -274,11 +279,14 @@ public class TeamController {
 	
 	// 칸반 받기
 	@RequestMapping(value = "cooperation-kanban.do", method = RequestMethod.GET)
-	public String kanbanView(int tseq, Model model) {
+	public String kanbanView(int tseq, Model model,HttpSession session) {
 		String path = "";
+		String email = (String)session.getAttribute("email");
 		ProjectDao dao = sqlsession.getMapper(ProjectDao.class);
 		List<Tissue> tissuelist = teamservice.loadKanban(tseq);
 		TeamPjt pjt = dao.detailPJT(tseq);
+		int rank = dao.searchRank(tseq, email);
+		model.addAttribute("rank", rank);
 		model.addAttribute("tpj",pjt); //프로젝트 이름 , 설명
 		if(tissuelist.isEmpty()) {
 			path = "cooperation/cooperation-kanban";
@@ -384,8 +392,8 @@ public class TeamController {
 	public Process chart(int tseq) {
 		
 		Process processList = null;
-		TissueDao dao = sqlsession.getMapper(TissueDao.class);
 		
+		TissueDao dao = sqlsession.getMapper(TissueDao.class);
 		
 		processList = dao.chartData(tseq);
 		
@@ -397,7 +405,7 @@ public class TeamController {
 	}
 	
 	@RequestMapping(value = "addTeamCalendar.do", method = RequestMethod.POST)
-	public String addTeamCalendar(HttpSession session,String title, String start, String end, String description, String type, String username, String backgroundColor, String textColor, boolean allDay, int tseq) {
+	public String addTeamCalendar(HttpSession session,String title, String start, String end, String description, String type, String username, String backgroundColor, String textColor, String allDay, int tseq) {
 		int result = 0;
 		System.out.println(title+"/"+start+"/"+end+"/"+description+"/"+type+"/"+username+"/"+allDay+"/"+tseq);
 		String viewpage = "";
@@ -410,7 +418,13 @@ public class TeamController {
 		tissue.setTseq(tseq);
 		tissue.setBackgroundColor(backgroundColor);
 		tissue.setTextColor(textColor);
-		tissue.setAllDay((true ? 1 : 0));
+		int alldayReturn = 0;
+		if(allDay.equals("true")) {
+			alldayReturn = 1;
+		} else {
+			alldayReturn = 0;
+		}
+		tissue.setAllDay(alldayReturn);
 		if(start.length()==16) {
 			System.out.println(start+":00");
 			tissue.setTistart(java.sql.Timestamp.valueOf(start+":00"));
@@ -436,17 +450,17 @@ public class TeamController {
 	}
 	
 	@RequestMapping(value = "editTeamCalendar.do", method = RequestMethod.POST)
-	public String editTeamCalendar(String title, String start, String end, String description, String type, String backgroundColor, boolean allDay, int tseq, int tiseq) {
+	public String editTeamCalendar(int _id, String title, String start, String end, String description, String type, String backgroundColor, boolean allDay) {
 		int result = 0;
-		System.out.println(title+"/"+start+"/"+end+"/"+description+"/"+type+"/"+allDay+"/"+tseq);
+		System.out.println(title+"/"+start+"/"+end+"/"+description+"/"+type+"/"+allDay+"/");
 		String viewpage = "";
 		Tissue tissue = new Tissue();
 		System.out.println(start.length());
+		System.out.println("idididididididididididid :"+_id);
 		MyIssueDao myissuedao = sqlsession.getMapper(MyIssueDao.class);
 		tissue.setTititle(title);
 		tissue.setTicontent(description);
-		tissue.setTseq(tseq);
-		tissue.setTiseq(tiseq);
+		tissue.setTiseq(_id);
 		tissue.setBackgroundColor(backgroundColor);
 		if(start.length()==16) {
 			System.out.println(start+":00");
@@ -475,20 +489,26 @@ public class TeamController {
 	}
 	
 	@RequestMapping(value = "deleteTeamCalendar.do", method = RequestMethod.POST)
-	public String deleteTeamCalendar(int tiseq) {
+	public String deleteTeamCalendar(int tiseq, String username, HttpSession session) {
 		int result = 0;
 		String viewpage = "";
 		Tissue tissue = new Tissue();
 		MyIssueDao myissuedao = sqlsession.getMapper(MyIssueDao.class);
 		tissue.setTiseq(tiseq);
-		result = myissuedao.deleteTeamCalendar(tissue);
-		
-		if(result>0) {
-			System.out.println("성공");
-			viewpage = "redirect:/calendar.do";
-		} else {
-			System.out.println("실패");
-			viewpage = "redirect:/calendar.do";
+		try {
+			String name = (String)session.getAttribute("name");
+			if(name.equals(username)) {
+				result = myissuedao.deleteTeamCalendar(tissue);
+				if(result>0) {
+					System.out.println("성공");
+					viewpage = "redirect:/calendar.do";
+				} else {
+					System.out.println("실패");
+					viewpage = "redirect:/calendar.do";
+				}
+			}
+		} catch (Exception e) {
+			// TODO: handle exception
 		}
 		
 		return viewpage;
@@ -544,7 +564,7 @@ public class TeamController {
 				JSONObject data = new JSONObject();
 				int z = sortlist.get(key).getAllDay();
 				
-				data.put("_id", key++);
+				data.put("_id", sortlist.get(key).getTiseq());
 				data.put("title", sortlist.get(key).getTititle());
 				data.put("description", sortlist.get(key).getTicontent());
 				boolean allDay;
@@ -611,6 +631,22 @@ public class TeamController {
 	public String ladder(int tseq,Model model) {
 		
 		return "user/projectLadder";
+	}
+	
+	@RequestMapping("/calendar.do")
+	public String object(@RequestParam(required = false, name="lang") String language, HttpServletRequest request, HttpServletResponse response,Model model,HttpSession session) {
+		System.out.println("calendar 왔냐?");
+		if(language == null) {
+			language = "ko";
+		}
+		String email = (String)session.getAttribute("email");
+		ProjectDao dao = sqlsession.getMapper(ProjectDao.class);
+		List<Tpmember> mem = dao.getTpMember(email);
+		Locale locale  = new Locale(language);
+		System.out.println("locale : " + locale + "\n language : " + language);
+		localeResolver.setLocale(request, response, locale);
+		model.addAttribute("mem",mem);
+		return "sidebar/calendar";
 	}
 	
 }
